Moved Platform class to separate module and made it a singleton.
