+++ doc/Glossary.rst
@@ -6,20 +6,36 @@ Glossary
    Base-Class
      An ancestor class for other derived classes.
 
+   Cygwin
+     ...
+
+   Decorator
+     ...
+
    Exception
-     A ...
+     ...
 
    Meta-Class
-     A ...
+     ...
 
    MinGW
      Minimalistic GNU for Windows.
 
    MSYS2
+     ...
 
    native
      A *native environment* is a platform just with the operating system. There is no additional environment layer like
      MSYS2.
 
+   Overloading
+     ...
+
+   Singleton
+     ...
+
    UCRT
      Universal C Runtime
+
+   WSL
+     Windows System for Linux
